Add uci.Engine.Getpid()
This change adds a Getpid() method receiver to uci.Engine. This
enables callers to interact with the engine process directly. For
example callers may renice the process to allow the engine to run
at a lower priority. In practice this is helpful in improving system
responsiveness when running the engine with high thread counts and/or
larger hash sizes.