Add uci.Engine.Renice()
This change adds a Renice() method receiver to uci.Engine. This
enables callers to specify that the engine should be run at a lower
priority. In practice this is helpful in improving system
responsiveness when running the engine with
high thread counts and/or larger hash sizes.

(cherry picked from commit 92c19038af6ce77a985384e1081b920345cd417c)