Added multiple functions related to Moves.
Move.GetTags() allows for external code to query the tags to store in a database.
Move.MovedPiece() allows for external code to query what piece was moved in a move
Game.MoveSquares() allows external code to initiate a movement between two Squares
Game.GetMove() allows for external code to query a specific move in the history of the game without copying the array.
NewSquareFromStr() allows for external code to convert strings into Squares

Updated code in the Engine and Decoding and Square to support these functions
Added new tests and updated others to incorporate new functions