Add position.XFENString()
When working with 3rd party chess APIs these may require positions to
be expressed in X-FEN notation rather than FEN. e.g. Notably while
some of the lichess.org APIs work with either FEN or X-FEN, some of
them require X-FEN. This commit adds position.XFENString() for this
purpose. The key difference between X-FEN and FEN is the encoding of
the en passant square. X-FEN will only specify it when an opposing
pawn is in position to capture, while FEN will always specify it.