Fix PGN parsing of games w/ nested variations (#125)
* Fix PGN parsing of games w/ nested variations

Existing moveListWithComments() contains a parsing bug when the pgn in
question contains a game with nested variations. In this case the
moveListTokenRe regex doesn't work as expected and results in attempts
to parse an illegal game. To demonstrate the problem this commit adds
a TestScannerWithNested() test case along with a 0013.pgn fixture
which is an export of a lichess study. This commit also fixes the
problem by adding a stripVariations() preprocess step in
moveListWithComments(). I have a subsequent commit coming which adds a
Scanner option to include variations in the game list when processing
a pgn.

* simplified tests

---------

Co-authored-by: Logan Spears <loganjspears@gmail.com>